Update rhel.Dockerfile
while there's no point in using micro here as we're not actually publishing this image, we can switch to use it for the Brew bootstrap runtime image. 

In a future PR, I may merge the Dockerfile, brew.Dockerfile and rhel.Dockerfile into a single file, so that we have less to maintain here, and as in other projects, transform it with minimal ssed changes so it works in Jenkins to build the arch-specific binaries, and then in Brew using those tarball'd binaries.